+++ TARGETS
@@ -18,21 +18,21 @@ cpp_library_wrapper(name="rocksdb_lib", srcs=[
         "cache/fast_lru_cache.cc",
         "cache/lru_cache.cc",
         "cache/sharded_cache.cc",
+        "cloud/aws/aws_file_system.cc",
         "cloud/aws/aws_kafka.cc",
         "cloud/aws/aws_kinesis.cc",
         "cloud/aws/aws_retry.cc",
         "cloud/aws/aws_s3.cc",
-        "cloud/aws/fs_aws.cc",
         "cloud/cloud_env_options.cc",
         "cloud/cloud_file_cache.cc",
         "cloud/cloud_file_deletion_scheduler.cc",
+        "cloud/cloud_file_system.cc",
+        "cloud/cloud_file_system_impl.cc",
         "cloud/cloud_log_controller.cc",
         "cloud/cloud_manifest.cc",
         "cloud/cloud_scheduler.cc",
         "cloud/cloud_storage_provider.cc",
         "cloud/db_cloud_impl.cc",
-        "cloud/fs_cloud.cc",
-        "cloud/fs_cloud_impl.cc",
         "cloud/manifest_reader.cc",
         "cloud/purge.cc",
         "db/arena_wrapped_db_iter.cc",
@@ -365,21 +365,21 @@ cpp_library_wrapper(name="rocksdb_whole_archive_lib", srcs=[
         "cache/fast_lru_cache.cc",
         "cache/lru_cache.cc",
         "cache/sharded_cache.cc",
+        "cloud/aws/aws_file_system.cc",
         "cloud/aws/aws_kafka.cc",
         "cloud/aws/aws_kinesis.cc",
         "cloud/aws/aws_retry.cc",
         "cloud/aws/aws_s3.cc",
-        "cloud/aws/fs_aws.cc",
         "cloud/cloud_env_options.cc",
         "cloud/cloud_file_cache.cc",
         "cloud/cloud_file_deletion_scheduler.cc",
+        "cloud/cloud_file_system.cc",
+        "cloud/cloud_file_system_impl.cc",
         "cloud/cloud_log_controller.cc",
         "cloud/cloud_manifest.cc",
         "cloud/cloud_scheduler.cc",
         "cloud/cloud_storage_provider.cc",
         "cloud/db_cloud_impl.cc",
-        "cloud/fs_cloud.cc",
-        "cloud/fs_cloud_impl.cc",
         "cloud/manifest_reader.cc",
         "cloud/purge.cc",
         "db/arena_wrapped_db_iter.cc",
@@ -4938,6 +4938,12 @@ cpp_unittest_wrapper(name="clipping_iterator_test",
             extra_compiler_flags=[])
 
 
+cpp_unittest_wrapper(name="cloud_file_system_test",
+            srcs=["cloud/cloud_file_system_test.cc"],
+            deps=[":rocksdb_test_lib"],
+            extra_compiler_flags=[])
+
+
 cpp_unittest_wrapper(name="cloud_manifest_test",
             srcs=["cloud/cloud_manifest_test.cc"],
             deps=[":rocksdb_test_lib"],
@@ -5428,12 +5434,6 @@ cpp_unittest_wrapper(name="filename_test",
 
 cpp_unittest_wrapper(name="flush_job_test",
             srcs=["db/flush_job_test.cc"],
-            deps=[":rocksdb_test_lib"],
-            extra_compiler_flags=[])
-
-
-cpp_unittest_wrapper(name="fs_cloud_test",
-            srcs=["cloud/fs_cloud_test.cc"],
             deps=[":rocksdb_test_lib"],
             extra_compiler_flags=[])
