+++ cloud/cloud_file_system_test.cc
