+++ cloud/cloud_file_system_wrapper.h
@@ -4,7 +4,7 @@
 #include <atomic>
 #include <thread>
 
-#include "cloud/fs_cloud_impl.h"
+#include "cloud/cloud_file_system_impl.h"
 #include "rocksdb/cloud/cloud_env_options.h"
 #include "rocksdb/cloud/cloud_storage_provider.h"
 #include "rocksdb/status.h"
