+++ cloud/db_cloud_impl.cc
@@ -5,8 +5,8 @@
 
 #include <cinttypes>
 
+#include "cloud/cloud_file_system_impl.h"
 #include "cloud/filename.h"
-#include "cloud/fs_cloud_impl.h"
 #include "cloud/manifest_reader.h"
 #include "env/composite_env_wrapper.h"
 #include "file/file_util.h"
