+++ cloud/db_cloud_test.cc
@@ -12,11 +12,11 @@
 #include <cinttypes>
 
 #include "cloud/cloud_file_deletion_scheduler.h"
+#include "cloud/cloud_file_system_impl.h"
 #include "cloud/cloud_scheduler.h"
 #include "cloud/cloud_storage_provider_impl.h"
 #include "cloud/db_cloud_impl.h"
 #include "cloud/filename.h"
-#include "cloud/fs_cloud_impl.h"
 #include "cloud/manifest_reader.h"
 #include "db/db_impl/db_impl.h"
 #include "db/db_test_util.h"
