+++ cloud/manifest_reader.cc
@@ -5,11 +5,11 @@
 
 #include <unordered_map>
 
+#include "cloud/cloud_file_system_impl.h"
 #include "cloud/cloud_manifest.h"
 #include "cloud/cloud_storage_provider_impl.h"
 #include "cloud/db_cloud_impl.h"
 #include "cloud/filename.h"
-#include "cloud/fs_cloud_impl.h"
 #include "db/version_set.h"
 #include "env/composite_env_wrapper.h"
 #include "rocksdb/db.h"
